+++ packages/ledger-icrc/src/ledger.canister.spec.ts
@@ -1,7 +1,7 @@
 import type { ActorSubclass } from "@dfinity/agent";
 import { Principal } from "@dfinity/principal";
 import { arrayOfNumberToUint8Array } from "@dfinity/utils";
-import { mock } from "jest-mock-extended";
+import { mock } from "vitest-mock-extended";
 import type {
   Allowance,
   ApproveArgs,