+++ packages/sns/src/governance.canister.spec.ts
@@ -5,7 +5,7 @@ import {
   InvalidPercentageError,
   toNullable,
 } from "@dfinity/utils";
-import { mock } from "jest-mock-extended";
+import { mock } from "vitest-mock-extended";
 import type {
   ListNervousSystemFunctionsResponse,
   ListProposalsResponse,